feat: migrate MovementSystem test as unit test example
- Moved MovementSystem.test.ts to be adjacent to MovementSystem.ts source file
- Updated imports to use centralized test utilities from tests/helpers/
- Applied standardized ECS test environment setup
- Improved test naming for clarity (removed console.log, better descriptions)
- Removed duplicate test from __tests__/ directory after successful migration
- Focuses on isolated unit testing with proper dependency mocking

Part of #46 testing structure reorganization
T004 completed: exemplary unit test migration pattern established
